pyproject.toml
[project]
name = "ai-research-notes"
version = "0.1.0"
description = "Multi-topic ML/AI research and learning repository"
requires-python = ">=3.12"
dependencies = [
    "cma>=4.4.2",
    "gymnasium[box2d]>=1.2.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
